Creating the User Register Form template
First of all, I had to create the view called RegisterView. This class bassed-view handles the get request creating a form
and sending it to the template through the context. The post request logic fills the form with the information in the
POST request, then checks if the form is valid, saves the user and flash messages. I can get the username from the
form cleaned data dictionary (this dict appears from the is_valid method call) in order to send it in the flash message.

VIEW Flash message:
1. messages.type:
	messages.success(request, "string with the message")
BASE TEMPLATE FLASH MESSAGE:
1. Django sends the messages for free through the "messages" context variable. This is a list with all the flass messages
2. Iterate through the list. {%for message in messages%}
3. Create a div with the class "alert alert-{{message.tags}}" and put the {{message}} in the div's content

FORM HTML
1. Create the form tags
2. Fieldset with the class "form-group"
3. Legend with "border-bottom mb-4"
4. Use Crispy tag
5. Sign up button
6. Login text