Creating the logout template and adding more tags to the login template
In order to display on screen a logout template when the user clicks on the logout anchor tag and it doesn't have
a next parameter in the url, I had to create a template view that renders the 'registration/logout.html' template.
By default, Django sends the user to the logout template from the admin site but I don't want the user to
look at this syte, I want the user to have another template. So, I edited the LOGOUT_REDIRECT_URL in settings.py.

If I want to redirect the user to a wished url (Django sends the user to 'accounts/profile' by default), I also had
to edit LOGIN_REDIRECT_URL