Editing the base template
I added a bootstrap nav bar and the hrefs with the url Django template tag, depending on the user state that is checked
with "{% user.is_authenticated %}". If that is true, the navbar shows the Logout and the Profile anchor tags, otherwise,
it shows the Login and Register anchor tags.

I also used a Bootstrap row with two colums in order to place the flash messages and the content in the first column and
a sidebar in the second column