Creating the update and delete mark view
In order to create the items mor responsive, I used the class="d-flex flex-row justify-content-between align-items-center"
and the id="Check-Item-{{ task.id }}" in a div that contains each list element. The id is used to get the element in the
future with a JS function.

Each checkbox has the onclick="markTask('{% url 'tasks:mark' task.id %}', {{ task.id }})". This is used to pass the url and
the id in order to send a post request to that url with AJAX and update the database without refreshing the page. The function
gets the label tag and checks if this tag has the line-through, then it updates that style. The Django View gets the object
according to its primary key and updates the task state.

The delete button <button class="btn btn-outline-danger" onclick="deleteTask('{% url 'tasks:delete' task.id %}', {{ task.id }})
uses the same parameters in the funtion in order to send the post request and deletes the task from the database and
remove the list item from the current page