Creating a Tag in order to give the user the capacity to group tasks
MODEL:

1. The tag has a model that contains the tag title and the tag author. Then, the way to connect the Tag and Task model
is through a ManyToMany Field given that a tag can contain multiple tasks and a task can have multiple tags.

That's the why of the ' tag = models.ManyToManyField('Tag', related_name="all_tasks") ' field in the task model. Related
Name parameter is used to grab all the tasks that are part of a tag

FORM:
The Tag form is a simple Model Form provided by Django. But the Task Form it's a little bit different cuz I had to override
the __init__ method, because I had to make a query to grab all the tags created by the current logged in user, and the default
ModelForm displays on screen all the tags, no mather who the tag author is. So, in the view I Passed the request.user object
to the form class constructor, like thid:
form= TaskForm(user=request.user)

And this parameter arrives to the self method in the '**kwargs' arguments. In order to take the user from the the **kwargs,
I used the pop method to grab the 'user' object like this:
user = kwargs.pop('user')

Finally, the tag field is a ModelMultipleChoiceField and this has the queryset parameter, so I had to pass the Queryset filtering
by the user. 'user.all_tags.all()' gives me all the tags created by the user.