Adding the edit tag color feature
MODEL: The model has two charfields, the first one is used to store the CSS class name and the other one is used to
store a human readable color in order to display that on screen

TEMPLATE: The templates have the {{ tag.color.color_class }} in the class attribute in order to give some style to
the tag element and {{ color.color_name }} in the tag list view to put the edit color options in a dropdown. Code:

      <ul class="dropdown-menu">
      {% for color in colors %}
        <li>
          <a class="dropdown-item d-flex" href="{% url 'tasks:edit_color' tag.id color.id %}">
            <div class="bg-{{ color.color_class }} rounded-circle"></div>
            <div class="mx-2">{{ color }}</div>
          </a>
        </li>
      {% endfor %}
      </ul>

The url pattern used to change the tag color has two integers passed as parameters in the url. That are the tag id
and the color id.

VIEW: The view gets the tag object and the color object from the model, then, it updates the color field in the tag record
and redirects the user to the tag list view