Creating the update task feature
This feature allows the users to update their tasks when they click on the update button. This button is gonna pass (to its
corresponding JS function) the url where the app wants to make the HTTP POST request and the task id. In the JS file, it
has been declared the postURL and labelID already, so everytime the user clicks on the edit task button, that values are gonna
be overwritten.

Then, if the user clicks on the "Save changes" button that appears with the modal, the event listenner added on the modal will
take the input text and put the value in the HTML label tag that shows the task. It also will make a POST request through the
fetch API to the url passing the text data. The second argument of the fetch call is an object with the request information, such
as HTTP method, headers (because I have to pass the csrf token in all the posts requests), the data in the body propertie as
a JSON string (I passed an object with the key of "new_title" and its value is the updated tag text.

It's important to know that the "Accept": "application/json" propertie has to be in the headers because that tells Django
the data passed in the post request is in the JSON format.

VIEW:
The view is pretty simple, first of all, it gets the model object with the id passed through the url and reads the request object
with json.load(request) and it sets the task title field to the data passed by the POST request. Then, it returns a response object