Creating the Update and Delete tag Features
ROUTING:
I created two url patterns that get the tag id and pass it to their corresponding views in order to get a tag object
and update it in the database. Code:

    path('update/tag/<int:pk>', TagUpdateView.as_view(), name='tag_update'),
    path('delete/tag/<int:pk>', TagDeleteView.as_view(), name='tag_delete'),

UPDATE VIEW:
This view renders the tag model form in the tag_form.html template in the HTTP GET request. Then, it validates the form
in the post() method and updates the record in the database if the form is valid, otherwise, it will render the
template again with its errors.

DELETE VIEW:
The delete view is going to render a tag_delete.html template asking the users if they want to delete the tag item. If
the user accepts, it deletes the record from the database and redirects the user to the main tasks list template.